💚CI : Remove Dependency-Submission
Gradle doesn't currently provide a mechanism to determine the location where a dependency is declared.
In fact, the resulting dependency version can be influenced by many different sources within a Gradle project.
So in my case, since gradlew is not located in root, there is no way to run it.